move master-slave spring namespace related to shardingsphere-master-slave-spring-namespace
